refactor(테마커뮤니티): 각 컴포넌트 요소들 분리 #17 - SearchBar, ActivityTab, KeywordTab, Tabmenu 경로 분리 - src/components/community/* 와 src/components/community/tab/* - 각 컴포넌트들의 export는 각 경로의 index.ts로 관리하여 하나의 경로로 모두 불러올 수 있게 설계
