+++ src/components/community/index.ts
@@ -1,0 +1,5 @@
+export * from './tab';
+
+export { default as CommentModal } from './CommentModal';
+export { default as CommunityPostGridItem } from './CommunityPostGridItem';
+export { default as SearchBar } from './SearchBar';
