+++ src/components/community/tab/KeywordTab.tsx
@@ -1,0 +1,6 @@
+export default function KeywordTab() {
+  return (
+    <>
+    </>
+  );
+}
