+++ src/components/community/tab/index.ts
@@ -1,0 +1,3 @@
+export { default as ActivityTab } from './ActivityTab';
+export { default as KeywordTab } from './KeywordTab';
+export { default as TabMenu } from './TabMenu';
